+++ src/suitsTerminal/Interfaces/IUnlockableSuit.cs
@@ -1,4 +1,5 @@
 ﻿using InjectionLibrary.Attributes;
+using suitsTerminal.Suits;
 
 [assembly: RequiresInjections]
 
@@ -8,8 +9,10 @@ namespace suitsTerminal.Interfaces;
 public interface IUnlockableSuit
 {
     //lets us create a patch for spawn of the suit
+    [HandleErrors(InjectionLibrary.ErrorHandlingStrategy.Ignore)] //it's okay if someone else injects this method
     void Start();
 
     //suitsTerminal attributes
-    SuitAttributes Attributes { get; set; }
+    [HandleErrors(InjectionLibrary.ErrorHandlingStrategy.LogWarning)] //should never happen but who knows
+    SuitAttributes SuitsTerminalAttributes { get; set; }
 }